do not allow to delete trainings in the past

+++ app/Http/Controllers/TrainingController.php
@@ -81,6 +81,7 @@ final class TrainingController extends Controller
                 'id' => $training->id,
                 'name' => $training->name,
                 'place' => $training->place,
+                'start_at' => $training->start_at,
                 'start_at_day' => $training->start_at->format('Y-m-d'),
                 'start_at_time' => $training->start_at->format('H:i'),
                 'length' => $training->length,
@@ -105,6 +106,11 @@ final class TrainingController extends Controller
 
     public function destroy(TrainingDestroyRequest $request, Training $training)
     {
+        // do not allow to delete a training in the past
+        if ($training->start_at < now()) {
+            return Redirect::back()->with('error', 'Múltbeli edzés már nem törölhető!');
+        }
+
         $training->delete();
 
         return Redirect::route('trainings')->with('success', 'Edzés sikeresen törölve.');
